# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "INZBC API",
    "summary": "Trade intelligence and FTA services for the India New Zealand Business Council.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/candidates": {
      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Capture Candidate",
        "operationId": "capture_candidate_api_candidates_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CaptureCandidateIn"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "List Candidates",
        "operationId": "list_candidates_api_candidates_get",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "run",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Run id to list candidates for.",
              "title": "Run"
            },
            "description": "Run id to list candidates for."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CandidateOut"
                  },
                  "title": "Response List Candidates Api Candidates Get"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}": {
      "get": {
        "tags": [
          "Candidates"
        ],
        "summary": "Get Candidate",
        "operationId": "get_candidate_api_candidates__candidate_id__get",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/verify": {
      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Verify Candidate",
        "operationId": "verify_candidate_api_candidates__candidate_id__verify_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/score": {
      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Score Candidate",
        "operationId": "score_candidate_api_candidates__candidate_id__score_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/route": {
      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Route Candidate",
        "operationId": "route_candidate_api_candidates__candidate_id__route_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/candidates/{candidate_id}/merge": {
      "post": {
        "tags": [
          "Candidates"
        ],
        "summary": "Merge Candidate",
        "operationId": "merge_candidate_api_candidates__candidate_id__merge_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "candidate_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Candidate Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MergeIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidateOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CandidateOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "captured_at": {
            "type": "string",
            "title": "Captured At"
          },
          "in_coverage_window": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Coverage Window"
          },
          "nz_relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nz Relevance"
          },
          "india_relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "India Relevance"
          },
          "member_relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Relevance"
          },
          "signal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Signal"
          },
          "confidence": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence"
          },
          "verification": {
            "type": "string",
            "title": "Verification"
          },
          "duplicate_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duplicate Of"
          },
          "included": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "proposed_routing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Routing"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_id",
          "headline",
          "source_id",
          "url",
          "summary",
          "published_at",
          "captured_at",
          "in_coverage_window",
          "nz_relevance",
          "india_relevance",
          "member_relevance",
          "signal",
          "confidence",
          "verification",
          "duplicate_of",
          "included",
          "reason",
          "proposed_routing"
        ],
        "title": "CandidateOut"
      },
      "CaptureCandidateIn": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "headline": {
            "type": "string",
            "title": "Headline"
          },
          "source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "in_coverage_window": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "In Coverage Window"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_id",
          "headline"
        ],
        "title": "CaptureCandidateIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "MergeIn": {
        "properties": {
          "duplicate_of": {
            "type": "string",
            "title": "Duplicate Of"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "duplicate_of",
          "reason"
        ],
        "title": "MergeIn"
      },
      "RouteIn": {
        "properties": {
          "proposed_routing": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Proposed Routing"
          },
          "included": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Included"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "RouteIn"
      },
...
      "ScoreIn": {
        "properties": {
          "nz_relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nz Relevance"
          },
          "india_relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "India Relevance"
          },
          "member_relevance": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Relevance"
          },
          "signal": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SignalStrength"
              },
              {
                "type": "null"
              }
            ]
          },
          "confidence": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SourceConfidence"
              },
              {
                "type": "null"
              }
            ]
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "reason"
        ],
        "title": "ScoreIn"
      },
...
      "SignalStrength": {
        "type": "string",
        "enum": [
          "Low",
          "Medium",
          "High",
          "Critical"
        ],
        "title": "SignalStrength"
      },
      "SourceConfidence": {
        "type": "string",
        "enum": [
          "High",
          "Medium",
          "Low",
          "Unverified"
        ],
        "title": "SourceConfidence"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "VerificationState": {
        "type": "string",
        "enum": [
          "Verified",
          "Partially Verified",
          "Unverified",
          "Not Required",
          "Rejected"
        ],
        "title": "VerificationState"
      },
      "VerifyIn": {
        "properties": {
          "verification": {
            "$ref": "#/components/schemas/VerificationState"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "verification",
          "reason"
        ],
        "title": "VerifyIn"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Opaque server-side session, issued out of band. HttpOnly, Secure, SameSite=Lax and host-only, so it is not readable by script and does not cross an origin. Only its SHA-256 digest is stored.",
        "in": "cookie",
        "name": "inzbc_session"
      },
      "CsrfToken": {
        "type": "apiKey",
        "description": "Double-submit token, returned by GET /api/session. Required on every state-changing request: SameSite=Lax still permits a top-level cross-site POST, which is the shape of a form-submission CSRF.",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  }
}